+++ CHANGELOG.md
@@ -2,13 +2,52 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+<a name="0.6.3"></a>
+## [0.6.3](https://github.com/EricLambrecht/better-spotify-playlists/compare/v0.6.2...v0.6.3) (2018-08-28)
+
+
+### Miscellaneous
+
+* **ui:** Removed default spotify URI from input.
+
+
+<a name="0.6.2"></a>
+## [0.6.2](https://github.com/EricLambrecht/better-spotify-playlists/compare/v0.6.1...v0.6.2) (2018-08-28)
+
+
+### Bug Fixes
+
+* **ui:** Prevent playlist image aspect ratio from changing ([5b49f93](https://github.com/EricLambrecht/better-spotify-playlists/commit/5b49f93))
+
+
+
+<a name="0.6.1"></a>
+## [0.6.1](https://github.com/EricLambrecht/better-spotify-playlists/compare/v0.6.0...v0.6.1) (2018-08-28)
+
+
+### Bug Fixes
+
+* **ui:** Fixed error message not disappearing. Also removed tracks in case of error. ([2ba9c69](https://github.com/EricLambrecht/better-spotify-playlists/commit/2ba9c69))
+
+
+
+<a name="0.6.0"></a>
+# [0.6.0](https://github.com/EricLambrecht/better-spotify-playlists/compare/v0.5.0...v0.6.0) (2018-08-28)
+
+
+### Features
+
+* **ui:** Added a playlist dropdown for easier playlist selection ([13cb061](https://github.com/EricLambrecht/better-spotify-playlists/commit/13cb061))
+
+
+
 <a name="0.5.0"></a>
 # [0.5.0](https://github.com/EricLambrecht/better-spotify-playlists/compare/v0.4.1...v0.5.0) (2018-08-24)
 
 
 ### Features
 
-* **playlist-selector:** Error messages are now shown to the user ([2c3a279](https://github.com/EricLambrecht/better-spotify-playlists/commit/2c3a279))
+* **ui:** Error messages are now shown to the user ([2c3a279](https://github.com/EricLambrecht/better-spotify-playlists/commit/2c3a279))
 
 
 
@@ -23,25 +62,25 @@ All notable changes to this project will be documented in this file. See [standa
 
 
 <a name="0.4.0"></a>
-## 0.4.0 (2017-04-04)
+# 0.4.0 (2017-04-04)
 ### Features
 * Added support for more than 100 tracks per playlist
 
 <a name="0.3.0"></a>
-## 0.3.0 (2017-04-03)
+# 0.3.0 (2017-04-03)
 ### Features
 * Added ability to define playlist start time.
 * Fixed several bugs
 * Refactored code base
 
 <a name="0.2.0"></a>
-## 0.2.0 (2017-04-03)
+# 0.2.0 (2017-04-03)
 ### Features
 * Adjusted icon size
 * fixed callback url (to work on servers other than "localhost:8080").
 
 <a name="0.1.0"></a>
-## 0.1.0 (2017-04-03)
+# 0.1.0 (2017-04-03)
 
 ### Features
 * First working draft.